ok 6

+++ WS/J02_introduccion2/src/ejercicios/ejercicios06/Ej06.java
@@ -6,7 +6,7 @@ public class Ej06 {
 		
 	}
 	
-	public static long potencia (int num, int exponent) {
+	public static double potencia (int num, int exponent) {
 		if (exponent == 0){
 			return Math.pow(num, exponent = 1);
 		}
